Add files via upload
This project is a personal job portal website designed and developed by Akash Tanwar as part of the Web Development Task assigned by Novanectar Services Pvt. Ltd. The objective was to create a professional and visually appealing platform to showcase personal skills, projects, and resume in a company-style layout.

The website serves as a digital resume and job portfolio, presenting Akash as a potential candidate for internships and jobs. It is structured like a company career portal to give a formal and modern appearance, suitable for industry visibility and recruitment.